Server: Add node registration logic
Create the first layout of the node registration logic in the server.
This will add a node to a list within the server when
/adap_parameter/Register is called, so long as the node that calls it
also advertises ~/Tune.

The logic will replace the node in the list on subsiquent calls to
Register, and nodes that have died (~/Tune has been lost), will also be
replaced by newly registered nodes.

The Requirements Specification.md has beed updated to better reflect
these changes.