graph.py: Automatically display parameter graph
Spruce up build.py and add graph.py to automatically build and run a
node/server and display a graph of the parameters tuned (or other info
from the client). The information should be printed to stdout from the
client in CSV form with the first row being the headers.