Add Canny client
On the LTU ThinkPad cameras, this does not require any alpha tuning in
the server to work flawlessly. Note that the "ratio" for canny is very
low, only 1.1 (where 1 is minimum), since that allows the tuning server
to have more feedback. Higher ratios make the tuning steps harder to
jump to.